Extracted some utils methods.
