+++ test/e2e/support/utils/get-json-response.js
@@ -1,0 +1,12 @@
+/**
+ * Get a JSON response for the passed in object, for use with `request.respond`.
+ *
+ * @param {Object} obj Object to seralise for response.
+ * @return {Object} Response for use with `request.respond`.
+ */
+export function getJSONResponse( obj ) {
+	return {
+		content: 'application/json',
+		body: JSON.stringify( obj ),
+	};
+}
